Fix issue for "reposition" calculation on re-open
During the "create" flow the modals height and margin are properly
calculated.  However, any time after that the modal height is evaluated
as zero because height is evaluated before the "display:none" is removed
from the element.  Changing the location of the "jQuery.show()" command
remedies this situation and the modal height and margin are correct and
consistent.