Update pgwmodal.js
Consider adding CSS class 'pgwModalOpen' also to the html tag element and avoid "page scroll" when html selector has overflow-y property declared. In that case adding pgwModalOpen to the body tag element doesn't seem to be enough in order to avoid scrolling the page below the backdrop element. Not sure about ('body, html') as the best way of declaring the selectors and targeting the elements but you get the gist of it.